backend/requirements.txt
Django==3.2
django-cors-headers==3.13.0
uwsgi==2.0.17.1
djangorestframework==3.14.0
PyJWT==1.7.1
dj-rest-auth==2.2.5
django-allauth==0.51.0
Pillow==9.0.1
djangorestframework-simplejwt==5.2.2
# djoser==2.1.0
psycopg2==2.9.3
